add #792 - Number of Matching Subsequence
This showed up in a youtube video of a mock interview, so I tried to find my own solution but it ended up being really similar to what the interviewee came up with (stringSubsets2), so thats what it morphed into. One problem is that each map value in this solution is an array (instead of a stack)... which adds overhead in having to find and remove items. I'm still kinda murky on the order that the items are entering and exiting the array - so I specifically found the appropriate index value in the arrays and spliced them out of it .... I think using a stack or deque to push/pop would clear up that murkiness.
Then I realized that a much simpler to understand solution (stringSubsets) is to loop through the words, filter the string for the letters in each word, and count it as a result if the filtered s value includes the full word