add #221 - Maximal Square
Found this in a Google mock interview youtube video and worked through the recursion solution along with it (eh_farm_square files) .. but I set up the matrix with number values.
Turns out when I found 221 on leetcode, I realized the values are strings... oops. So I when ahead and copied in the recursive solution there... then tried figuring out the DP table solution which was not so great. This is a bit over my head right now, so I copied in a better 1D DP solution ... and I'm using this to work toward a better understanding